Prepare platform for 386 merge; merge 386_39179 + binary blobs (RT-AC86U)
- Reorganized all prebuild_arm_* folders to use model-based paths
  (like rc prebuild).  Symlinks are used for common files.
  This is necessary as the use of the EX7 and EXHND variables was
  non-consistent between mainline and ax, causing conflicts.
  Makefiles will now always use prebuild/$(BUILD_NAME), with
  symlinks handling the rest.
- Merged GPL 386_39179 + binary blobs from RT-AC86U
- Updated rc/openvpn.c and httpd for OpenSSL 1.1 compatibility
- Removed kludges for 382 models
- Rewritten networkmap traffic meters for new style
- Harmonized OpenSSL and inadyn code with upstream
- Re-synced rstats code with upstream (hopefully the
  random rate spikes are fixed) since it carries
  many important enhancements for LACP and others.

+++ release/src/router/libupnp_arm_94908hnd/upnp/upnp_gena.c
@@ -244,7 +244,7 @@ notify_prop_change(UPNP_CONTEXT *context, UPNP_SUBSCRIBER *subscriber)
 
 	/* set non-blocking IO to have connect() return without pending */
 	on = 1;
-	if ((ioctl(s, FIONBIO, (int)&on)) == -1) {
+	if ((ioctl(s, FIONBIO, (long)&on)) == -1) {
 		upnp_syslog(LOG_ERR, "cannot set socket as non-blocking IO");
 		goto error_out;
 	}
@@ -287,7 +287,7 @@ notify_prop_change(UPNP_CONTEXT *context, UPNP_SUBSCRIBER *subscriber)
 
 	/* set non-blocking IO to have connect() return without pending */
 	on = 0;
-	if ((ioctl(s, FIONBIO, (int)&on)) == -1) {
+	if ((ioctl(s, FIONBIO, (long)&on)) == -1) {
 		upnp_syslog(LOG_ERR, "cannot set socket as blocking IO");
 		goto error_out;
 	}
